refactor: useImageInput 훅을 renderHook 패턴으로 리팩터링
post api 요청을 react query로 분리
