+++ frontend/src/apis/oAuthApi/index.ts
@@ -4,6 +4,7 @@ import {
   getTokenGoogle,
   patchMyInfo,
   deleteMyInfo,
+  postProfileImage,
 } from 'apis/oAuthApi/api';
 import {
   GetMyInfoResponse,
@@ -72,3 +73,13 @@ export const useDeleteMyInfo = (
     deleteMyInfo,
     options,
   );
+
+export const usePostProfileImage = (
+  options?: UseMutationOptions<AxiosResponse, AxiosError<ErrorResponse>, FormData>,
+) => {
+  return useMutation<AxiosResponse, AxiosError<ErrorResponse>, FormData>(
+    'postProfileImage',
+    postProfileImage,
+    options,
+  );
+};
