+++ frontend/src/mocks/handlers/auth.js
@@ -50,7 +50,8 @@ export const auth = [
     return res(ctx.status(204), ctx.delay(1000));
   }),
 
-  rest.post('/images', (req, res, ctx) => {
+  // 프로필 이미지 업로드(POST)
+  rest.post(`${BASE_URL}/members/me/profile-image`, (req, res, ctx) => {
     console.dir('@@@프로필 이미지 post 성공', JSON.stringify(req.body));
     return res(ctx.status(201));
   }),
